Recreated projects targeted for SL4.  Couldn't get SL5 beta installed.
I really want the async/await functionality of the Async Ctp, and it
just isn't possible to do with SL5, so I've recreated the projects
targeted now to SL4.  It at least compiles.

Also, I didn't realize I didn't have the .gitignore, so I've added that
and I'll need to clean up the files already in place that should have been
ignored.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>