Total Revamp Architecture to attempt to properly configure Csla dataportal.
I didn't realize it, but until this point I had been developing Csla with
only 2-tier architecture, and thus was not comprehending the full
awesomeness that is the data portal.  It abstracts away the channel
between the client and server, so that the WCF service that is run is
not my own service, but a Csla dataportal service, WcfPortal.svc and
SlPortal.svc.  I learned this as I finally was able to get the
Csla Configuring Data Portal ebook.

So, I redid the architecture to where it at least compiles. The projects
are currently as follows:
LearnLanguages(.)
-.Business.Client
-.Business.Server
-.Common
-.Common.Server
-.DataAccess
-.DataAccess.Client
-.IISHost
-.Silverlight

Now, I think I understand that the host WCF service is actually executed
by the IIS environment.  I had to add the IIS web page, add two services,
WcfPortal.svc and SlPortal.svc, and point them properly to the Csla
data portal interface (Csla.Server.Hosts.WcfPortal and
Csla.Server.Hosts.Silverlight.WcfPortal respectively).  Now the IIS
Express runs and hosts these services at
http://localhost:50094/WcfPortal.svc and
http://localhost:50094/SlPortal.svc (port 50094 on my machine).  So the IIS
environment is the process that is actually running on the server and it
in turn is the thing that actually exposes these services.

Also, I'm cleaning up dependencies, trying to expose as little as the
DataAccess layer to the client as necessary, and understanding better
which methods are called on the server versus the client.  I had figured
a lot of this out a long time ago, but that's one of the drawbacks of
coding rarely, you forget a lot in between.  Anyway, I'm much more
refreshed with Csla and becoming more and more comfortable with it.

I do not have any test projects at the moment, but I will soon.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>