Restructured Dal to more closely mimic Csla recommendations/samples.
-Now the DalManager is created by DalFactory.GetDalManager(), using
appSettings in Web.config.  IDalManager is in DataAccess, and the
DalManager implementation is in each concrete provider project.

The individual IPhraseDal implementations
are then accessed via DalManager.GetProvider<IPhraseDal>() which uses a
naming convention.  This assumes that the implementation is called
"PhraseDal", specifically that it is IPhraseDal minus the 'I'.  The
convention also assumes that the PhraseDal is in the same namespace as
the DalManager implementation.

-Reformatted all the DataPortal calls in PhraseEdit and LanguageEdit
to reflect this change. NB that NEITHER of these now use Mef/DI to resolve
the IPhraseDal/ILanguageDal interfaces.

-Added LearnLanguages.DataAccess.Mock, but not filled out with actual
MockDb.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>