Wcf talking now at least. Finally!!!! Basic Compression Implemented.
In my attempt to getting it working, one of the suggested solutions was
using compression for silverlight, which I planned on anyway, so that is
implemented now, using additional library suggest: SharpZipLib.  The license
says works in closed commercial apps.

The errors I had to get through were:
1) I had to not only rename my
business output libraries to be the same across Client and Server,
ie (LearnLanguages.Business in both .Business.Client and .Business.Server
project properties), but also manually "clean" solution by deleting bin
and obj folders.
2) I had to make sure that my server had references to both csla and my
server-side business library.

I went ahead and built a sample solution that mimicked the IISHost sample
in the Csla DataPortal Configuration downloads.  (I will be putting
that up here as well.)  I went through and checked every reference and used
their SlPortal.svc and WcfPortal.svc verbatim.

-ALSO, I'VE ADDED A CLIENTACCESSPOLICY.XML FILE TO ELIMINATE CROSS-DOMAIN
PROBLEMS THAT I'VE READ ABOUT.  This was another possible solution, and
I will be taking this out to see if it works without it.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>