Made two tests for large and varied texts assigned to Phrases.
Saves and Loads just fine.  Tested up to 500,000 character string, which
has no problems persisting and being retrieved.  Test is set to 30,000
characters for speed.  500,000 on my machine takes about 15 seconds or so,
but that time is generating the string, not the actual db calls.  Those
are very fast.

Also, I rechecked the necessity for the dummy class in the all-link
Business.Server project, and it does seem necessary.  Apparently it needs
one solid, non-link class file in order to compile the dll properly.
I did set the dummy to internal class.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>