Problems with Static LearnLanguagesContextManager class.
This thing will hit breakpoints on first run of tests.  Then, when
the test fails (LanguageTests.GET_ALL()), if I start the same test,
with no changes in code, the breakpoints will not be hit.  It will
just throw an exception when the static class is accessed.  I've had
problems with static classes as singletons before, and I am going to
switch this over to a singleton in the Csla.ApplicationContext.LocalContext
to see if this helps.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>