Oh Wow.  LanguageTests are passing with EF!!!! OMG!!! Finally!
Apparently I am correctly seeding the EF database now, and updating
the SeedData as well.  It definitely helped to turn the static classes
into full, multi-threaded singleton instances.  That removed a lot of
doubt and weird quirky misfirings of static constructors and what-have-you.
Anyway, going on to the PhraseTests now.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>