Implementing CustomIdentityDal in Ef.  Passing some security tests.
Instead of implementing ICustomIdentityDal directly, I've created a
CustomIdentityDalBase class in .DataAccess that wraps calls to abstract
protected implementation methods, e.g. VerifyUserImpl(), GetUserImpl().
This way, I can have the try..catch..throw_specific_exception behavior
in the base class and share this for any other .DataAccess.XXX
providers (Ef in this case).  See .DataAccess.CustomIdentityDalBase and
.DataAccess.Ef.CustomIdentityDal.  I will be changing over my other Dal's
to do this as well.  It's a much more consistent and neat structure.

I also made the convention that for any Dal method, there will be an
exception in the form of [DalMethod]FailedException, and in DalResources
there will be an error message ErrorMsg[DalMethod]FailedException.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>