Fixed DB doubling glitch.Understand seeding inconsistency.Working on Tests.
-PhraseList was not loading child phrases properly.  Fixed this by changing
DataPortal<PhraseEdit>.ChildCreate(dto) to .ChildFetch(dto), propagated
changes.

-Seeding inconsistency comes from the project not being rebuilt every
single time.  So, the static instances in between debugging executions
are not reset to null.  Literally, if I start the debug session then end
it and immediately hit F5 again, then when I check a singleton's value
immediately it will not be null!! Cleaning _any_ project, explicitly
stopping IIS server, or hitting a rebuild all will "fix" this.
I tried setting a pre/post build event to kill the IIS process, but
this event does not get triggered because the solution does not get built!

-PhraseTests are passing now.  Divided PhraseTests and PhraseListTests.
PhraseLists are not working at moment.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>