All Tests Passing.
-I think I'm not doing something quite right with my declarations because
I should just be able to say List.BeginEdit(), List[0].Language = blahblah,
List.ApplyEdit(), but it is giving me a mismatch.  For now, I am able to
set the language and apply the edit to the individual PhraseEdit, just
not the List.  I can save the list, just not applyedit in this way that I
expected I would.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>