Trying to get AddTranslations to work properly, but failing.  Moving on.
-I still don't have the mechanism quite right for AddTranslations.  It
does not want to include the default language properly on the
TranslationPhrasesItemViewModel's LanguageSelector.  Basically, it works,
but if you add a phrase, then try to save without manually selecting its
language, then it will throw an exception.  I'm moving on however, as I
don't see this as a viable/practical way to add translations anyway.
That will be done through the learning process by selecting text and making
translations out of the selections, which btw I looked up google translate
API and that is a no-go, as it costs money.  Microsoft's Bing Translate
API is free however, so I am going to try to incorporate that into the
program.  It's pretty much 100% MS anyway, so I think that should be fine.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>