Ef provider working now with translations.
-Functionality is working great as far as I can tell at the moment with
the Ef provider.  I'm able to view translations, edit translations,
edit phrases.  When I delete a phrase, it automatically deletes the
translation if that translation has fewer than two phrases.
-Tests do not all pass at once, but this is because of my evidently
poorly structured testing, not the functionality.  I will look at this
when I think it's necessary/when I think I'll gain something from
correcting it.  (It looks like it would be quite an investment as it is a
fundamental flaw in my testing process.  I would have to seed each test
class and only affect those records that I seed initially.  I'm not sure
that the rigor is worth it.)  In the meantime, I will think about how
to do testing more cleverly in the future.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>