Updated README: Notes on building/running and Journal.
-I wanted to give instructions as follows:
+NOTES ON BUILDING/RUNNING:
+-Be sure to manually set the IISHost project as the StartUp Project.
+-To switch between the Mock and Ef data providers: IISHost project -> Web.config -> <appsettings> -> key="DalManagerType".
+-To switch between Silverlight Unit Testing UI and the actual Silverlight app UI: IISHost project -> Project Properties -> Web tab -> Specific Page(.TestsTestPage is the UnitTesting page right now - bad naming sorry).

-Journal:
+01/29/12
+I've integrated TranslationEdit functionality into both Mock and Ef
Data Providers now.  A translation is a collection of phrases, requiring
 at minimum two phrases.  In the UI, I can now add translations, view
translations, edit translations, but not delete them (the functionality
exists but just not in UI yet).  If deleting a _phrase_ causes a
translation to contain only one or zero phrases, however, then the
translation _is_ automatically deleted cascadingly.  I cannot add
existing phrases to a translation yet.  Overall, making progress with
basic functionality & plumbing, and almost to the point of the cooler
stuff with the learning process.  Nothing is optimized for speed, all
for functionality, so that is also probably on the horizon once I'm
working with actual larger data sets.  I also anticipate wanting to
download a backup of data file.  All in all going well.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>