Navigation pulled out into its own project. InProgress: IStudyPartner
-Navigation is a little better, abstracted into its own project.  Still
needs work, but working as well as it was before.
-Changed "home" page to StudyViewModel.
-Started on NavigationSets, for a better way to navigate/organize
navigation.

-Added basic, cheap loginfailed handling.  Not really secure against a
brute force attack that modifies client code I don't think.
-Working on Study Partner stuff.  Implementing a basic CycleStudyPartner.
This will use extra data (CycleStudyDataEdit or something similar) to
to make decisions on what to study.  Right now, I'm thinking only of
collecting the user's native language.  Which reminds me..

-Need to remove user's native language from database and put it in the
CycleStudyData's database.  This should be separate I think, not
directly associated with User or Language Tables, but I will have to think
about this.  I'm tired but current thinking is: Any study information
should be thought of as self-contained, "viewing" the study process on its
own.  So, we are using our current data access layer because of convenience
but ideally a study partner is a pluggable entity that can handle its own
data.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>