Lots of changes.  Jeeez.
-Added LanguageDal.Fetch(languageText);
-Added PhraseDal.New(languageText).  Actually this is adjusting
already used .New(object criteria) and checking if criteria is string.
With if it is string, then it is assumed to be the languageText, and
creates the new PhraseEdit loading it with that language.  Propagated
changes necessary for this.
-Moved IStudyPartner and IAskUserExtraData to Silverlight project. Not
sure if I should move to pluggable architecture for resolution of
IStudyPartner or just have my default study partners in the Silverlight
project and worry about a more pluggable dynamic composition later.
I'm leaning towards the former.
-Created TranslationRetriever(phrases) to make it easier to create
a translation with given phrases as children.  The phrases may start
out as objects not marked as children, so this copies them to
newly added children of Translation.Phrases.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>