Add Song working for the most part.  InProgress: UI stuff.
-Add Song now takes in song lyrics text, parses, creates phrases for all
words, Lines/Phrases for each Line, and a Song that holds these LineEdits.
It persists all of these in the datastore (in a single transaction I think)
-Now I'm working on Navigation better to use navigation sets and to
"finalize" or at least more concretize my UI experience.  I'm currently
thinking of Navigation Sets, and how this would interact with a
navigation panel.
-Also, I went ahead and added ability for RandomPhraseStudyPartner to
study in other languages than english/spanish.  If the question is english,
it assumes you want to study the spanish phrase.  If the question
however is in italian or german, then it will automatically make the answer
english (my native language) and autotranslate it with the correct codes
for Bing auto translation.  This sometimes is causing some kind of terrible
crash however, and I think this has to do with the Bing translator service.
I need to wrap this in a custom translator interface with bing as one
implementation of the auto translate interface (service).

Signed-off-by: William Raiford <bill.mybiz@gmail.com>