Lot of UI consistency stuff.  AskUserExtraData is working ftmp. Bugs.
-Instructions: One instructions per viewmodel.  thinking about putting in
IViewModel or ViewModelBase for standardized instructions on the
view.  Less cluttered.  Using tooltip bindings for individual instructions.
-AskUserExtraData is working now.  If no StudyDataEdit is found for
the current user, it prompts the user to enter it.  Right now, that is
only the user's native language, but it is good because it ties the
NativeLanguageText to our data via the language selector.  It does not
have strict referential integrity on purpose.
-Bugs in AddSong view model.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>