InProgress: Exchange pattern.
-Having to focus on distributed execution more closely.  Expounding on
offer pattern with a more generalized "Exchange pattern".  This is
producer-consumer-esque.  Instead of a work queue static in memory though,
the opportunities are produced temporally in real-time, and either
someone is listening (listener pattern) or not.  This is more than
just a listener pattern however, as the listeners engage in a dialog
of Opportunity -> Offer -> OfferResponse, in a strictly decoupled fashion.
They communicate through interfaces (IOpportunity, IOffer, IOfferResponse).
Status will be optional through the exchange as well.
-I'm reusing my old statuses from Common, adding ErrorEncountered,
Progressed, and Other.
-The workflow will be as follows:
1) Exchange.SubscribeToOpportunities/Offers(this) (which for now will
just relay to a now internal IEventAggregator, and will assume
the caller knows to use IHandle<msg>.  In future, may need to abstract this
to my own event aggregating, or move to something like NServiceBus).
2) Publish opportunities based on events (e.g. user clicks on study song,
song is selected, job opportunity is created, job is published).
3) Worker hears opportunity, checks out if it can do it, publishes offer.
4) Job Poster hears offer, decides if job is still available and if
it is not superceded by something else.  Publishes Accept/Deny.
5) Worker does work, publishing status using exchange as well.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>