InProgress: Exchange/Study.  It's building though!
-Breaking to figure out how I want the product of the exchange to happen.
I was envisaging the studiers to just publish navigation events, but
I'm not liking this, especially since the UI project already references
the study project.  And I have the NavigationController's method for
locating viewmodels to be looking in Silverlight.ViewModels.  I'm not sure
if I want to change this.  I am thinking I will have a product produced
by the Exchange doer.  So, it would go:
Opportunity > Offer > OfferResponse > OpportunityCompleted; with the
OpportunityCompleted having a reference to all of the steps in the
process, as well as the Opportunity.Fruit produced by the work.  It would
then consume this fruit.  In my current usage, this would be a ViewModel
of type IViewModelBase that would be assigned to a property in the calling
ViewModel, e.g. StudyASongViewModel."CurrentPromptViewModel".  As long
as the ViewModel and its View are in the container, I think that Caliburn
will be able to find them and inject the view into the original
StudyASongView <ContentControl> element.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>