InProgress: Exchange/Study. Added a .Product to IJobInfo<TTarget, TProduct>
-Decent compromise.  Declaring looks terrible because of all of the
generic type parameters everywhere, but it does the trick.  Essentially,
along with the job info when the job is created, the publisher has an
optional product type that it expects.  This is where whoever does the job
puts the object that they job was intended to create.  So, I put in
IViewModelBase as my intended product, assign this Product when I need it
(in DefaultPhraseMeaningStudier I believe), and then I publish the
StatusUpdate.  Now the publisher has the finished product that it shows
to the user for however long _it_ wants.
-This raises the question of who decides how long to show what, but there
should be a way for either party to put timers on how they use this
viewmodel.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>