A QA ViewModel has been shown!  Studiers have produced something tangible.
-Finally!
-It's not even close to working, but I finally was able to get a
viewmodel inserted in the StudyASongViewModel, and got Caliburn to
be able to inject its view (even though the view is in a different
assembly).  It doesn't translate, it doesn't allow for translation,
and it throws exceptions if you try to do anything more than once,
but it still shows something!  Whew!
-Now I need to change a bunch of the code, as the messages are being
thrown around like crazy.  I need to change the bubbling up events
to internal conglomerate messages, as I think that will be neater.
-I still don't really like calling studier.Do to initiate behavior
and then listening to exchange for result, but I think I will fix that when
I have to.  For now, it is doable so to speak.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>