Some kinks worked out.  Still totally flawed.
-Need to figure out how to better encapsulate "saga" information.
-Currently, the VM publishes an opportunity with job information.
Handshaking ensues: Offer supplied, offer responded to...So far so good.
But here is the problem.  I was thinking of something like a yield
statement, where you Do(job). Stop. Do(job). Stop.  But how to accomplish
this exactly is the details in the devil.  I think I need to rethink
what the study partner is trying to do exactly, now that we have
the product being produced.  We now have control over what ViewModel
is shown in the main window (albeit indirect control).  We can
create our own study view model that gets shown inside of the invoking
ViewModel.  We should be able to just "Go" with the study partner,
and not worry about the exchange.  I'm going to need to rethink how/why
I have disconnected my study partner pieces as I have.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>