-Compiling and sort of Q&Aing!
-Have one iteration of study working, in that it will show a Q & A,
with the same phrase as question and answer, for a limited amount of time.
-It doesn't go on to the next item.  It doesn't do a lot of things.  But,
it what it does do I am somewhat more pleased on how it does it.  The
studier components now each have a simpler async base class with a method
that initializes and one that performs a study iteration, using callback
instead of pub-sub.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>