Modifying study still.
-Was working much better, cycling through lines, and aggregated phrases.
-Still no persistent knowledge however, so does not ignore known things.
-Still no feedback.  Started on this, with creating feedback view/viewmodel.
-Now restructuring what StudyASongViewModel does with resulting viewmodel.
Currently, this is to create a StudyViewModel, which is just a housing
ViewModel for the one created by the study partner.  StudyASongViewModel
creates this, sets its StudyScreen to the one gotten from the exchange,
and then publishes a navigation request.
-Created Home View/ViewModel for when having logged in.  This empty at the
moment but in the future will be news/updates, learning status.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>