Studier sort of working with background worker thread.
-I still don't like it, but it seems to be doing basically what it is
supposed to do.  It has a "nested" call in the background worker
completed event handler, that invokes the background worker to
run async.  This is how it iterates through study items.  I don't really
know if it is working, because the view is messed up, but I need to go
and don't have time to go to that aspect at the moment.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>