Fixed the views to where you can see the study item and feedback.
-The process is moving along.  you can now click some feedback and it will
go to the next item to study.  It still has no knowledge of feedback,
so it will cycle through all of the items indefinitely, but it is still
somewhat working.
-Deleted the unused coroutines folder.  That was good practice though if
I end up needing them in UI stuff.  Also, good thinking on IEnumerable.
I may end up using something with IEnumerable anyway since I consider
the study steps to be iterations.  For now, I'm going with the
background worker thread though.
-Starting on abstracting the creation of the study item view model.  I
want to have ability to have various study item view models as an
extensibility point, so I am not stuck with just a simple Q and A item.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>