Refactored Retrievers/Creators. Added context to translation.
-Some retrievers I thought of more as creators, I refactored.
-Added ContextPhrase property to TranslationEdit.  Added this to
edmx as well, with a 0..1 to many relationship with phrases.
So now, a translation can have a context phrase, which does indeed
mean that there is a many to many relationship between phrases
and translations.
-Examples: "the" and "der" has a
context of "masculine form".  "the" and "die" with a context of
"feminine form".  "the" and "die" with a context of plural form.
"dog" and "Hund" has no context (context == null).  "blade" and
"rollerbladen" has a context of "skating".  "blade" and "Messer"
has a context of "cutting instrament".
-The context phrase must be a phrase saved in the database.  Notice
though that this can be in any language, so it can be in the native
language of a German speaker.  E.g. "cat" and "Katze" can have a
context of "animalisch".

Signed-off-by: William Raiford <bill.mybiz@gmail.com>