Working much better.  Still in progress distributed event system.
-Phrases are aggregating finally!  It sort of remembers what words/phrases
are "known"/"unknown".  However, the way it stands now it pumps out only
the phrase question, which some are in native language and some are
in the foreign language.  So the aggregation does not work properly here.
I need to change the saga information to include foreign/native language
information in the phrase events (oh no!), so the advisor can use
this information to know that both phrases are understood.  The other route
is to say that they are separate phrases and that this helps you learn the
song from both languages.  I don't like this though, so I'm leaning towards
adding information to the saga.  I'm not looking forward to it, but it's
nice to finally have some tangible progress and I'll try to focus on that
progress.
-Added ConsoleHistoryListener, that outputs to Debug.  Will refactor this
later to DebugHistoryListener.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>