Abort mechanism implemented.
-Each piece is responsible for listening for the event messages
that would require that piece to abort.  In this case, this means
that each studier listens for a NavigationRequested message.  I may
abstract this later to the study partner listening for the conditions
and then publishing an abort study event message, but for now I'm unsure
of the various pros and cons.
-The general pattern I used was to have a (locked property) AbortIsFlagged,
which checks for this are peppered throughout each piece's code.  Except
with the viewmodels, I have the code directly in the IHandle handler,
as these do not have async problems to deal with.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>