Got this bizzle up and goin again.
Had a slight hiccup with sql (connection string, db didn't exist).
I included a SQL problem-solution.txt file describing this.  Would be
better to have some sort of issue tracker, but this is good enough for
now.  Anyway, back up and lernin pretty quick: 1 day (couple hours).  That
includes environment and everything.
-Reminds me...if you have a problem with the unit testing projects,
be sure to install the silverlight 4 toolkit, that has the unit testing
framework in it.

-Bill