Gott im Himmel jeez trying to get this editing working properly.
-Tired, think I'm messing up more than I'm fixing. The question and
answer PhraseEdits being supplied to the StudyViewModel are not the
ones in the DB.  Concretely, if I have a question in French "je suis"
and it auto translated (and auto saved) the translation with phraseedits:
{ id = 1, l=french, text="je suis" } & { id = 2, l=english, text=i am }
PhraseEdit with id 1 is getting passed to the viewmodel with a different
id like id 3.  So when I go to fetch the phrase edit from the DB by id,
it aint findin it and I apparently have not created any plumbing to fetch
a phraseEdit by its text property.  So I'm going round and round with
async nonsense (I hate that I couldn't use yield statements and shit
to get async better, but the ms library didn't work with the SL library,
so I have callbacks all over the place if I want to do any one little
thing...only other option is to make some UoW object for every little
which is also a pain in the ass to maintain).  Anyway, I'm going to step
away from this and look at it again tomorrow.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>