Still implementing Edit Translation stuff.
-Slept on it.  Decided to make phrase retrievers.
-First made one that
searches translations for a phrase pairing
(TranslationPairSearchRetriever/Criteria).  But then I realized that I was
needing this with a Phrase reviewer viewmodel, not a translation review
viewmodel.  So I didn't do any testing with these classes.
-So then I made Retriever that just gets PhraseEdits from DB given
a list of PhraseEdits as the criteria (ListOfPhrasesCriteria). Using this,
however, I've found that I need to rethink what I'm doing yet again.
-So, I was thinking that the autotranslate should autosave the translations
to the DB, as I didn't think that it would be a problem with the partial
phrases created when I do the segmentation of the lines for study.  But
now I'm not sure.  I need to write this out more carefully somewhere...I
think I'll do it in the wiki.
-I also was getting a problem with PhraseList.GetAll().  It was locking up
the tightloop User.ChildFetch<> when loading all of the PhraseEdits that
populate the PhraseList.  So, I took out the loading of the user object
in the PhraseEdit.LoadFromDtoBypassingPropertyChecksImpl method. This could
have more profound effects, but so far I haven't seen any.