Weee dogey that editing is somewhat done.
-Can edit the Question only now when studying a song, only using the
manual VM.
-Still need to add this same functionality to the Answer +& timed VM.
-The check for inserting/updating a phrase is now in the Ef PhraseDal
(not in the mock).
-Need to add a test for this in the PhraseTests.
-Still is a little slow, but I'm continuing not to worry about this though
it will be more and more important pretty soon.
-Along those same lines, I really need to have a consistent way of telling
the user that the program is thinking and disabling the entire view,
as well as a thinking in the background area.  Maybe a viewmodel in the
shell specifically devoted to this.  This is especially important, since
the future of this program will have this thing constantly thinking and
doing stuff (never just sitting there).

Signed-off-by: William Raiford <bill.mybiz@gmail.com>