Ahh fun. Trying to futz with updating knowledge before aggregating text.
-My program is choking on trying to GetPercentKnown() for lines.
-Introduced a timeout for a complicated async tight loop. Gonna have to
figure out how to do this right, cause right now it ain't. It takes about
1.5 minutes to initialize for a study session now, and that just isn't
right. And this mostly stems from my unweildy crappy async structure
that cascades Initialize for everything. I'll have to think on this awhile
to come up with what to do.
-Possibilities include: More unit testing for the advisor structure.
-Restructuring the study partner (as in another complete overhaul) in a
smarter way that anticipates long-running parallel processes. IOW,
when it initializes for study, what is the data that it really needs,
and just go for that at first. Then, in the background, prepare for
future studiers.
-Got to get a better grip on the long-running async pattern I think.
Just a couple of async processes I've done before, but this is a whole
new level of async that I need to manage.  Aww yeah.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>