I've gotten around the thread lock issue.
-Referring to the locking up when trying to do a tight loop using
ThreadPool and a WaitOne. Locking up the whole program. It was in
DefaultPhrasePercentKnownAdvisor. I had it looking up individual words
in a phrase in a tight loop, and this was for some reason blocking the main
thread. It's possible that I could ahve started it on a background worker
thread and this might fix the problem, but for now, it's not an essential
piece and so it just returns 0 pct known if it doesn't know the exact
phrase. It doesn't try to sum up the individual component words pct known.
-A note with usage. It does not form belief about aggregate phrase texts.
It only forms a belief (ie it only stores that it knows a phrase) when
that phrase is an entire line text. I had forgotten I made this decision.
I think it's a good design point, but I could change this in the future.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>