Working through AddUser for Ef Dal.
-Put CheckAuthentication() in CommonHelper instead of DalHelper.
-Also put the UserNotAuthenticatedException in the common project.
-Ef passing some AddUser tests!!
-Need to corect study data update behavior to check if need to
update vs need to insert.
-Messed around with some tests for ef add. So, I have a bunch of random
users in the ef database now, no big deal. I also have a user2/password2
user. I added a bunch of songs with that user. Seems to work ok.
-Can't test study a song yet, because study data native language stuff
doesn't work right. as I said up there, I need to correct the update
implementation for ef.
-Basically, just chugging through this AddUser stuff, and poking around
looking at behavior with exceptions.
-Need to have a better grip on these background worker threads going. If I
logout in the middle of one of them, it freaks out and it is pretty much
unrecoverable, as it will try to create viewmodels for all of the phrases,
for example, if that is what you were doing when you hit the logout.

Signed-off-by: William Raiford <bill.mybiz@gmail.com>